Make host-differential tests portable via stdint.h, override the rest
validate-unit-test.ps1 (host-gcc differential testing against the same
baselines runall.ps1 uses) had 10 failures. Investigated each:

tnarrow, tptrcst, tregnarw depended on dcc's 16-bit int (overflow wrap in
a running sum, or sizeof(int)*N pointer-diff scaling) which host's 32-bit
int can never reproduce with any compiler flag. Switched the specific
locals/arrays that actually need a fixed width to stdint.h's int16_t
(sizeof 2 on every platform, and identical to plain int on dcc's target,
so dcc's own generated code and cycle counts are unaffected) instead of
excluding these tests from host validation - now genuinely portable.

mm, tfdedge, tfmadd, tfmedge (dcc's software float doesn't bit-match host
hardware FPU rounding), tscanin (CP/M's blocking console read echoes
typed input; a host pipe doesn't) and tpragstk (tests the CP/M-only
stack-check guard, no host equivalent) can't be made portable this way,
so they're marked "host": true in _test_overrides.json.

tgnarly is mixed: its sizeof(int)-dependent local also got the int16_t
treatment, but it separately hits a trigraph divergence (gcc's -std=gnu99
disables trigraphs; C89 requires them) that isn't a dcc bug, so it's
still marked "host": true overall. It also surfaced a genuine dcc string-
literal bug, fixed separately.

Verified: runall.ps1 -Mode full -Extended -NarrowDiff all passing;
validate-unit-test.ps1 now 0 failures (was 10).

+++ tests/tptrcst.c
@@ -7,11 +7,15 @@
 
 #include <stdio.h>
 #include <stdlib.h>
+#include <stdint.h>
 
 typedef float ftype;
 
+/* iC/ip/ipend use int16_t (not plain int) so the expected byte diff (800,
+ * i.e. sizeof(int16_t)*400) is the same on dcc's 16-bit int and a host
+ * compiler's wider int. */
 ftype fC[20][20];
-int iC[20][20];
+int16_t iC[20][20];
 ftype fD[400];
 
 int main()
@@ -20,8 +24,8 @@ int main()
     ftype * fpend = ( (ftype *) fC ) + ( 20 * 20 );
     int fdiff = (int) fpend - (int) fp;
 
-    int * ip = (int *) iC;
-    int * ipend = ( (int *) iC ) + ( 20 * 20 );
+    int16_t * ip = (int16_t *) iC;
+    int16_t * ipend = ( (int16_t *) iC ) + ( 20 * 20 );
     int idiff = (int) ipend - (int) ip;
 
     ftype * dp = (ftype *) fD;
